backend/requirements.txt
fastapi
uvicorn[standard]
gunicorn
supabase
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
python-multipart
aiofiles
python-dotenv
